// types.ts
export type TransactionType = 'income' | 'expense' | 'transfer';
export type AccountType = 'bank' | 'cash' | 'person';
export type Currency = 'TL' | 'USD' | 'EUR' | 'TOMAN';
export type DateFormat = 'jalali' | 'gregorian';
export type AppTheme = 'light' | 'dark';
export type RecurrenceFrequency = 'daily' | 'weekly' | 'monthly' | 'yearly';

export interface SecuritySettings {
  enabled: boolean;
  pin: string | null;
  useBiometrics: boolean;
}

export interface AppSettings {
  appName: string;
  currency: Currency;
  dateFormat: DateFormat;
  theme: AppTheme;
  categories: {
    expense: string[];
    income: string[];
  };
  security: SecuritySettings;
}

export interface Account {
  id: string;
  name: string;
  type: AccountType;
  initialBalance: number;
  currency: Currency;
}

export interface Transaction {
  id: string;
  amount: number;
  targetAmount?: number;
  category: string;
  type: TransactionType;
  accountId: string; 
  toAccountId?: string; 
  date: string;
  note: string;
  isRecurring?: boolean; // Indicates if it was generated by a rule
}

export interface RecurringTransaction {
  id: string;
  amount: number;
  category: string;
  type: TransactionType;
  accountId: string;
  frequency: RecurrenceFrequency;
  startDate: string;
  nextDueDate: string;
  note: string;
  isActive: boolean;
}

export interface Summary {
  totalIncome: number;
  totalExpense: number;
  balance: number;
  balancesByCurrency: Record<Currency, number>;
}

export interface CategoryData {
  name: string;
  value: number;
}

export interface AccountSummary extends Account {
  balance: number;
}